move MutableBehavior out as separate file
