+++ akka-actor-typed-tests/src/test/scala/akka/actor/typed/WatchSpec.scala
@@ -6,7 +6,7 @@ package akka.actor.typed
 
 import akka.Done
 import akka.actor.typed.scaladsl.Behaviors
-import akka.actor.typed.scaladsl.Behaviors.MutableBehavior
+import akka.actor.typed.scaladsl.MutableBehavior
 import akka.actor.typed.scaladsl.adapter._
 import akka.testkit.EventFilter
 import akka.testkit.typed.scaladsl.{ ActorTestKit, TestProbe }
